requirements.txt
python-telegram-bot[job-queue]==20.8
openai==1.12.0
SQLAlchemy==2.0.25
APScheduler==3.10.4
psycopg2-binary==2.9.9
aiohttp==3.9.1
python-dotenv==1.0.0
